Questão 3 - Parte 2
A classe DataFrame foi separada em duas, com a nova subclasse DataFrameMath ficando responsável pelos métodos de cálculos. Com isso, a classe original fica responsável apenas pela construção e atualização do dataframe.
